(#11) Add functional testing
This patch adds functional testing for the scenario of an end user downloading
and installing the gem.  The goal is to make sure the gem installs cleanly and
basic commands and functionality work as expected.

Bash is used for functional testing to avoid issues related to bundler
polluting the runtime environment, e.g. as happened in
https://github.com/ghoneycutt/rizzo/issues/14

The testing environment is configured to be as clean as possible.  That is, no
other ruby gems other than the minimal set in the GEM_PATH.  This is why the
rvm gemset `cleanroom` is created.  Travis CI uses rvm, but local developers
like myself do not, so rvm is optional with `INSTALL_TO_GEM_HOME=true
./scripts/functional_gem_behavior.sh`

Tests cases covered are:

 * Install the gem
 * The gem and dependencies should be installed
 * The executable should be in the path
 * rzo --help should contain usage
 * rzo --version should output a semantic version string
 * rzo bare (no arguments) should match --help
 * rzo config with no config should be helpful
 * With a valid ~/.rizzo.json rzo config is expected to pretty generate the JSON config
 * rzo generate produces a minimal Vagrantfile
 * rzo generate STDERR is expected to match 'Wrote vagrant config to Vagrantfile'
 * rzo roles with no local .rizzo.json is expected to have no output
 * with a single puppetca role, rzo roles outputs the role name
 * rzo generate is expected to produce a Vagrantfile with one VM defined