(#16) Add json schema and directory existence validation
Without this patch the resulting Vagrantfile may be invalid, or rzo may be
configured to read from directories which do not exist.  This patch addresses
the problem by adding safety checking methods.

Json schema is also added to validate required keys in the configuration.  Two
passes are performed, one to verify the personal configuration file, typically
~/.rizzo.json, another pass to validate the full configuration merged with
.rizzo.json files in the top level repository.

Closes #16